core_calendar: Force index usage
On mysql-based db systems, the queries built by `raw_event_retrieval_strategy`
are supplemented by adding a `FORCE INDEX` statements between the `FROM`
clause and the `WHERE` clause.

This significantly improves performance on instances with a large number of
rows in the events table (we observed improvements from ~4m to ~500ms).

This patch has not been tested on PostgreSQL systems.

Co-authored-by: Stefan Auditor <stefan@auditor.email>
Co-authored-by: Andreas Grabs <moodle@grabs-edv.de>